Update App.js
In this i have used two hooks useState and useEffect .these two hooks are the most used hooks in react for dynamic content . we use useState for dynamic change of values and it accepts two parameter one is the current state and 2. is a function to update the state . useEffect is used when some sate is changed then what happened and on which which variable should we use it .in this we have again two parameters 1. function when our second parameter is changed then this annonymous function will  run 2. in this parameter we write array of dependencies.